+++ js/pages/monev.js
@@ -9,7 +9,7 @@ import {
   updateQueryString,
   announce
 } from "../app.js";
-import { getInitialData, refreshFromSheets } from "../data/data-service.js?v=0.9.4-deep-search";
+import { getInitialData, refreshFromSheets } from "../data/data-service.js?v=0.9.5-intent-search";
 import { emptyState } from "../ui.js";
 import { icon } from "../icons.js";